requirements.txt
streamlit>=1.28.0
numpy>=1.21.0
matplotlib>=3.5.0
twilio>=8.0.0
PyPDF2>=3.0.0
Pillow>=9.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
google>=3.0.0
openai>=1.0.0
pandas>=1.5.0
scikit-learn>=1.1.0
plotly>=5.0.0
qrcode>=7.3.0
opencv-python>=4.6.0
mediapipe>=0.9.0
pygame>=2.1.0
boto3>=1.26.0
streamlit-webrtc>=0.47.0
pyttsx3>=2.90
librosa>=0.10.0
psutil>=5.9.0
seaborn>=0.12.0
paramiko>=3.0.0
instagrapi>=2.0.0
